requirements.txt
discord.py==2.3.2
PyNaCl==1.5.0
yt-dlp==2024.3.10
ffmpeg-python==0.2.0
aiohttp==3.9.3

Flask==2.3.2
Flask-SQLAlchemy==3.0.3
requests==2.31.0
requests-oauthlib==1.3.1
waitress==3.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
python-dotenv==1.0.0

lyricsgenius==3.0.1
openai
googlesearch-python
youtube-search-python
deep-translator
fuzzywuzzy==0.18.0
webcolors
Pillow
